New SELinux domain for init.mmi.touch.sh
Create a new SELinux domain for init.mmi.touch.sh. Currently,
it's running in the shell SELinux domain, and can't perform the
actions it needs to perform. This hasn't been a problem yet,
since we haven't had to update firmware for the touch driver.

Don't put the shell script on the root partition. It's unneeded
and is better stored on /system where it can have a proper SELinux
label and be marked as executable.

Change-Id: I0902f7cec0e514e3af12db759c974de4f01a50f3